requirements.txt
django
djangorestframework
mysqlclient
gunicorn
python-dotenv
mangum